feat(checkin): Implemented checkin scanning system
Implemented a fully-functioning scanning system based on the admin dashboard.
Scan-in types are defined in the database and require a name and precedence for
sorting. The raw name of each scan-in determines whether or not a user is
checked-in to an event, has claimed merchandise, has claimed food, or attended a
workshop. Something that might need to be added is a check in the admin
dashboard to prevent non-admins/non-organizers from accessing the dashboard.
Currently, the only thing in the admin dashboard that does anything is the
scanner.

feat #6