Feat: Add Notion/markdown support and better sidebar to Hackerpack (#153)
* refactor: Cleaned up DocLinks component

* feat(hackerpack): Create dropdown menu for HackerPack mobile view

Add MobileDropdownMenu component that displays a mobile menu, used to navigate a
page using anchor tags. Also fix some erroneous CSS that broke the page
on a mobile view.

fix #147

* docs: Add tsdocs for MobileDropdownMenu component

Add component and usage information for the new MobileDropdownMenu component

* refactor: Clean up hackerpack mobile menu bar

Moved the MobileDropdownMenu component into the hackerpack folder. Also redid
the component so it no longer floats or obsures any page content. Added
animations to the dropdown and reworked the way the highlighting works.

fixes #147

* refactor: Add more sections to hackerpack sidebar example

Split the hackerpack sidebar sections into 3 sections to make the menu more
clear.

* feat: Set up framework for hackerpack customization

Created a settings JSON file for the hackerpack, as well as set up the framework
to implement the Notion native page render and markdown page rendering

* feat: Add Notion page "embed" to hackerpack

Added a native implementation of a Notion page embed that will pull data from
notion and format it to look like a Notion page. Also added server-side
rendering for the purposes of pulling the Notion data on the server side.

fix #149

* refactor: Change hackerpack from SSR to statically built

Changed the Notion data fetch to build time instead of server-side rendered on
each request, as the hackerpack shouldn't change too much (and we can just
rebuild to update). Also, the markdown conversion can happen in the same
getStaticProps function.

* feat: Add markdown support for hackerpack

Used react-markdown to load the hackerpack from a markdown file. Also added a
decleration module for `.md` files and a raw-loader extension to build `.md`
files.

fix #149

* feat: Add support for markdown tables and links

Added additional features to the markdown to html converter for the hackerpack.
This includes foramtted tables, links, strikethrough, bold, and italics.

* chore: Add a README for hackerpack customization

Created a README.md file for the hackerpack explaining the customization option
so organizers don't even have to touch the hackerpack code.

* refactor: Rename SidebarSection

Renamed MobileSidebarSection to simply SidebarSection as it's used in the
desktop sidebar too

* feat: Automatically generate hackerpack sidebar from markdown

Sidebar can now be automatically generated from the markdown file

* refactor: Reduce padding of hackerpack sidebar

Reduced the list padding due to it wrapping text on the desktop sidebar

* refactor: Remove sidebar for Notion pages

Due to Notion rendering random ids for the headings, the best option would be to
remove the sidebar/menu navigation if the user chooses to render a Notion page.
Notion has a native table of contents element for this purpose.

* fix: Remove weirdly spaced div on empty sidebar

A div that messed up the spacing of the hackerpack was returned when the sidebar
content was empty, this is now null

* feat: Add sidebar generation for Notion hackerpacks

Now the mobile menu + sidebar can be generated for Notion page embeds!

* docs: Update hackerpack sidebar generation on docs

Added info about generating the sidebar from Notion pages

* chore: Fix merge conflict with package.json

Fixed merge conflict due to multiple packages being added to develop branch and
feat-hackerpack, modifying the same lines in package.json

Co-authored-by: Abdullah H <71054105+abdullah50053@users.noreply.github.com>