fix(calendar): Fix `Schedule` error when no events (#223)
* fix(calendar): Only render track groups if there are tracks available

Since actualy track instances (like "General", etc) are computed from the
available appointments, the scheduler can't form groups if there are no
appointments available because it never expects there to be no instances of a
group when a group is specified. In reality, this is a bug in the scheduler
component itself, and should be fixed a different way, perhaps by patching React
Scheduler itself, but this will fix it for now by not attempting to render
groupings when there are no groupings available.

fixes #191

* fix(calendar): Move `GroupingState` out of conditional render

For some reason, `IntegratedGrouping` freaks out if GroupingState is not
rendered, but `GroupingState` doesn't cause any errors since it just contains
the empty groupings and doesn't try to render anything, so we can just render it
unconditionally and gate the buggy components behind the condition.

* fix(calendar): Render each grouping component in separate template brace

Since React Scheduler components rely on state of previously rendered
components, the only way I can see to solve this in a clean way is to render
each in a separate template brace. Not the clean solution I would've liked, but
it'll work and it's technically/logically correct.

---------

Co-authored-by: Abdullah H <71054105+abdullah50053@users.noreply.github.com>